feat: wallets tests

+++ backend/src/modules/wallets/wallets.service.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   ConflictException,
   Injectable,
   NotFoundException,
@@ -46,6 +47,12 @@ export class WalletsService {
         throw new NotFoundException('Carteira não encontrada');
       }
 
+      if (wallet.uuid !== user.uuid) {
+        throw new BadRequestException(
+          'Você não tem permissão para acessar esta carteira',
+        );
+      }
+
       return wallet;
     } catch (error) {
       throw error;
@@ -62,6 +69,12 @@ export class WalletsService {
 
       if (!wallet) {
         throw new NotFoundException('Carteira não encontrada');
+      }
+
+      if (wallet.uuid !== user.uuid) {
+        throw new BadRequestException(
+          'Você não tem permissão para atualizar esta carteira',
+        );
       }
 
       const walletByName = await this.walletsRepository.findByName(
@@ -85,6 +98,12 @@ export class WalletsService {
 
       if (!wallet) {
         throw new NotFoundException('Carteira não encontrada');
+      }
+
+      if (wallet.uuid !== user.uuid) {
+        throw new BadRequestException(
+          'Você não tem permissão para excluir esta carteira',
+        );
       }
 
       return await this.walletsRepository.delete(user.uuid, uuid);
